requirements.txt
python-telegram-bot[job-queue]==21.7
python-dotenv==1.0.1
openai==1.55.3
httpx==0.28.1
